requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
